Revert "Rename singer_sdk._singerlib -> singer_sdk.singer"
This reverts commit 14acaa42201f3b5b83486df18da761a794b4bde1.
