+++ .flake8
@@ -12,7 +12,7 @@ per-file-ignores =
     # Don't require docstrings conventions in "meta" code
     # singer_sdk/helpers/_classproperty.py:D105
     # Ignore unused imports in __init__.py files
-    singer_sdk/singer/__init__.py:F401
+    singer_sdk/_singerlib/__init__.py:F401
 max-complexity = 10
 docstring-convention = google
 allow-star-arg-any = true
