+++ singer_sdk/_singerlib/catalog.py
@@ -1,5 +1,3 @@
-"""Singer catalog parsing."""
-
 from __future__ import annotations
 
 import enum
@@ -7,7 +5,7 @@ import logging
 import typing as t
 from dataclasses import dataclass, fields
 
-from singer_sdk.singer.schema import Schema
+from singer_sdk._singerlib.schema import Schema
 
 if t.TYPE_CHECKING:
     from typing_extensions import TypeAlias
