+++ singer_sdk/_singerlib/schema.py
