+++ singer_sdk/helpers/_catalog.py
@@ -6,8 +6,8 @@ from typing import Any, Dict, Optional, Tuple
 
 from memoization import cached
 
+from singer_sdk._singerlib import Catalog, SelectionMask
 from singer_sdk.helpers._typing import is_object_type
-from singer_sdk.singer import Catalog, SelectionMask
 
 _MAX_LRU_CACHE = 500
