+++ singer_sdk/streams/rest.py
@@ -13,6 +13,7 @@ from warnings import warn
 import backoff
 import requests
 
+from singer_sdk._singerlib import Schema
 from singer_sdk.authenticators import APIAuthenticatorBase, SimpleAuthenticator
 from singer_sdk.exceptions import FatalAPIError, RetriableAPIError
 from singer_sdk.helpers.jsonpath import extract_jsonpath
@@ -23,7 +24,6 @@ from singer_sdk.pagination import (
     SimpleHeaderPaginator,
 )
 from singer_sdk.plugin_base import PluginBase as TapBaseClass
-from singer_sdk.singer import Schema
 from singer_sdk.streams.core import Stream
 
 DEFAULT_PAGE_SIZE = 1000
